refactor(bridge): move lazy-loaded components to a dedicated file for better organization
