+++ apps/cowswap-frontend/src/modules/bridge/index.ts
@@ -1,5 +1,3 @@
-import { lazy } from 'react'
-
 export { BridgeAccordionSummary } from './pure/BridgeAccordionSummary'
 export { useBridgeProvider } from './hooks/useBridgeProvider'
 export { useBridgeSupportedTokens } from './hooks/useBridgeSupportedTokens'
@@ -24,22 +22,5 @@ export {
 // Types
 export * from './types'
 
-// Lazy-loaded components for code-splitting
-// These will only be loaded when actually used in the application
-export const BridgeRouteBreakdown = lazy(() =>
-  import('./containers/BridgeRouteBreakdown').then((module) => ({
-    default: module.BridgeRouteBreakdown,
-  })),
-)
-
-export const BridgeStopDetails = lazy(() =>
-  import('./pure/BridgeStopDetails').then((module) => ({
-    default: module.BridgeStopDetails,
-  })),
-)
-
-export const SwapStopDetails = lazy(() =>
-  import('./pure/SwapStopDetails').then((module) => ({
-    default: module.SwapStopDetails,
-  })),
-)
+// Lazy-loaded components from dedicated file
+export { BridgeRouteBreakdown, BridgeStopDetails, SwapStopDetails } from './lazy'
